+++ PMC-Desktop/formEditUser.Designer.cs
@@ -23,6 +23,7 @@
         /// the contents of this method with the code editor.
         /// </summary>
         private void InitializeComponent () {
+            this.components = new System.ComponentModel.Container();
             this.buttonEUApplyChanges = new System.Windows.Forms.Button();
             this.labelEUFirstName = new System.Windows.Forms.Label();
             this.labelEUPrimaryEmailAddress = new System.Windows.Forms.Label();
@@ -43,13 +44,17 @@
             this.textEUManager = new System.Windows.Forms.TextBox();
             this.labelEUManagerName = new System.Windows.Forms.Label();
             this.labelEUUserEmailNotEnabled = new System.Windows.Forms.Label();
+            this.buttonEUCancel = new System.Windows.Forms.Button();
+            this.labelEUDisplayName = new System.Windows.Forms.Label();
+            this.textEUDisplayName = new System.Windows.Forms.TextBox();
+            this.errorEUNotify = new System.Windows.Forms.ErrorProvider(this.components);
+            ((System.ComponentModel.ISupportInitialize)(this.errorEUNotify)).BeginInit();
             this.SuspendLayout();
             // 
             // buttonEUApplyChanges
             // 
-            this.buttonEUApplyChanges.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this.buttonEUApplyChanges.DialogResult = System.Windows.Forms.DialogResult.OK;
-            this.buttonEUApplyChanges.Location = new System.Drawing.Point(315, 334);
+            this.buttonEUApplyChanges.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.buttonEUApplyChanges.Location = new System.Drawing.Point(315, 373);
             this.buttonEUApplyChanges.Name = "buttonEUApplyChanges";
             this.buttonEUApplyChanges.Size = new System.Drawing.Size(88, 34);
             this.buttonEUApplyChanges.TabIndex = 1;
@@ -68,8 +73,9 @@
             // 
             // labelEUPrimaryEmailAddress
             // 
+            this.labelEUPrimaryEmailAddress.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelEUPrimaryEmailAddress.AutoSize = true;
-            this.labelEUPrimaryEmailAddress.Location = new System.Drawing.Point(16, 206);
+            this.labelEUPrimaryEmailAddress.Location = new System.Drawing.Point(16, 245);
             this.labelEUPrimaryEmailAddress.Name = "labelEUPrimaryEmailAddress";
             this.labelEUPrimaryEmailAddress.Size = new System.Drawing.Size(110, 13);
             this.labelEUPrimaryEmailAddress.TabIndex = 3;
@@ -82,12 +88,13 @@
             this.textEUFirstName.Name = "textEUFirstName";
             this.textEUFirstName.Size = new System.Drawing.Size(130, 20);
             this.textEUFirstName.TabIndex = 4;
+            this.textEUFirstName.TextChanged += new System.EventHandler(this.textEUFirstName_TextChanged);
             // 
             // comboEUPrimaryEmailAddress
             // 
-            this.comboEUPrimaryEmailAddress.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.comboEUPrimaryEmailAddress.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.comboEUPrimaryEmailAddress.FormattingEnabled = true;
-            this.comboEUPrimaryEmailAddress.Location = new System.Drawing.Point(173, 203);
+            this.comboEUPrimaryEmailAddress.Location = new System.Drawing.Point(173, 242);
             this.comboEUPrimaryEmailAddress.Name = "comboEUPrimaryEmailAddress";
             this.comboEUPrimaryEmailAddress.Size = new System.Drawing.Size(230, 21);
             this.comboEUPrimaryEmailAddress.TabIndex = 5;
@@ -95,17 +102,18 @@
             // 
             // listEUProxyAddresses
             // 
-            this.listEUProxyAddresses.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.listEUProxyAddresses.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.listEUProxyAddresses.FormattingEnabled = true;
-            this.listEUProxyAddresses.Location = new System.Drawing.Point(173, 230);
+            this.listEUProxyAddresses.Location = new System.Drawing.Point(173, 269);
             this.listEUProxyAddresses.Name = "listEUProxyAddresses";
             this.listEUProxyAddresses.Size = new System.Drawing.Size(230, 95);
             this.listEUProxyAddresses.TabIndex = 6;
             // 
             // labelEUProxyAddresses
             // 
+            this.labelEUProxyAddresses.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelEUProxyAddresses.AutoSize = true;
-            this.labelEUProxyAddresses.Location = new System.Drawing.Point(19, 230);
+            this.labelEUProxyAddresses.Location = new System.Drawing.Point(19, 269);
             this.labelEUProxyAddresses.Name = "labelEUProxyAddresses";
             this.labelEUProxyAddresses.Size = new System.Drawing.Size(133, 13);
             this.labelEUProxyAddresses.TabIndex = 7;
@@ -113,7 +121,8 @@
             // 
             // buttonEUAddNewEmail
             // 
-            this.buttonEUAddNewEmail.Location = new System.Drawing.Point(77, 249);
+            this.buttonEUAddNewEmail.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.buttonEUAddNewEmail.Location = new System.Drawing.Point(77, 288);
             this.buttonEUAddNewEmail.Name = "buttonEUAddNewEmail";
             this.buttonEUAddNewEmail.Size = new System.Drawing.Size(75, 23);
             this.buttonEUAddNewEmail.TabIndex = 8;
@@ -128,6 +137,7 @@
             this.textEUMiddleName.Name = "textEUMiddleName";
             this.textEUMiddleName.Size = new System.Drawing.Size(107, 20);
             this.textEUMiddleName.TabIndex = 9;
+            this.textEUMiddleName.TextChanged += new System.EventHandler(this.textEUMiddleName_TextChanged);
             // 
             // textEULastName
             // 
@@ -136,6 +146,7 @@
             this.textEULastName.Name = "textEULastName";
             this.textEULastName.Size = new System.Drawing.Size(135, 20);
             this.textEULastName.TabIndex = 10;
+            this.textEULastName.TextChanged += new System.EventHandler(this.textEULastName_TextChanged);
             // 
             // labelEUMiddleName
             // 
@@ -158,7 +169,7 @@
             // labelEUDepartment
             // 
             this.labelEUDepartment.AutoSize = true;
-            this.labelEUDepartment.Location = new System.Drawing.Point(19, 70);
+            this.labelEUDepartment.Location = new System.Drawing.Point(19, 113);
             this.labelEUDepartment.Name = "labelEUDepartment";
             this.labelEUDepartment.Size = new System.Drawing.Size(62, 13);
             this.labelEUDepartment.TabIndex = 13;
@@ -181,15 +192,16 @@
             "Plumbing",
             "Sales",
             "Warehouse"});
-            this.comboEUDepartment.Location = new System.Drawing.Point(19, 86);
+            this.comboEUDepartment.Location = new System.Drawing.Point(19, 129);
             this.comboEUDepartment.Name = "comboEUDepartment";
             this.comboEUDepartment.Size = new System.Drawing.Size(243, 21);
             this.comboEUDepartment.TabIndex = 14;
+            this.comboEUDepartment.SelectedIndexChanged += new System.EventHandler(this.comboEUDepartment_SelectedIndexChanged);
             // 
             // labelEUTitle
             // 
             this.labelEUTitle.AutoSize = true;
-            this.labelEUTitle.Location = new System.Drawing.Point(19, 110);
+            this.labelEUTitle.Location = new System.Drawing.Point(19, 153);
             this.labelEUTitle.Name = "labelEUTitle";
             this.labelEUTitle.Size = new System.Drawing.Size(27, 13);
             this.labelEUTitle.TabIndex = 15;
@@ -197,15 +209,16 @@
             // 
             // textEUTitle
             // 
-            this.textEUTitle.Location = new System.Drawing.Point(19, 126);
+            this.textEUTitle.Location = new System.Drawing.Point(19, 169);
             this.textEUTitle.Name = "textEUTitle";
             this.textEUTitle.Size = new System.Drawing.Size(243, 20);
             this.textEUTitle.TabIndex = 16;
+            this.textEUTitle.TextChanged += new System.EventHandler(this.textEUTitle_TextChanged);
             // 
             // labelEUManager
             // 
             this.labelEUManager.AutoSize = true;
-            this.labelEUManager.Location = new System.Drawing.Point(19, 149);
+            this.labelEUManager.Location = new System.Drawing.Point(19, 192);
             this.labelEUManager.Name = "labelEUManager";
             this.labelEUManager.Size = new System.Drawing.Size(49, 13);
             this.labelEUManager.TabIndex = 17;
@@ -213,39 +226,76 @@
             // 
             // textEUManager
             // 
-            this.textEUManager.Location = new System.Drawing.Point(19, 166);
+            this.textEUManager.Location = new System.Drawing.Point(19, 209);
             this.textEUManager.Name = "textEUManager";
             this.textEUManager.Size = new System.Drawing.Size(243, 20);
             this.textEUManager.TabIndex = 18;
+            this.textEUManager.TextChanged += new System.EventHandler(this.textEUManager_TextChanged);
             // 
             // labelEUManagerName
             // 
             this.labelEUManagerName.AutoSize = true;
-            this.labelEUManagerName.Location = new System.Drawing.Point(268, 169);
+            this.labelEUManagerName.Location = new System.Drawing.Point(268, 212);
             this.labelEUManagerName.Name = "labelEUManagerName";
             this.labelEUManagerName.Size = new System.Drawing.Size(0, 13);
             this.labelEUManagerName.TabIndex = 19;
             // 
             // labelEUUserEmailNotEnabled
             // 
+            this.labelEUUserEmailNotEnabled.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelEUUserEmailNotEnabled.AutoSize = true;
             this.labelEUUserEmailNotEnabled.BackColor = System.Drawing.Color.Transparent;
             this.labelEUUserEmailNotEnabled.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.labelEUUserEmailNotEnabled.Font = new System.Drawing.Font("Microsoft Sans Serif", 26.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelEUUserEmailNotEnabled.Location = new System.Drawing.Point(15, 202);
+            this.labelEUUserEmailNotEnabled.Location = new System.Drawing.Point(15, 241);
             this.labelEUUserEmailNotEnabled.Name = "labelEUUserEmailNotEnabled";
             this.labelEUUserEmailNotEnabled.Size = new System.Drawing.Size(388, 41);
             this.labelEUUserEmailNotEnabled.TabIndex = 20;
             this.labelEUUserEmailNotEnabled.Text = "User Email Not Enabled";
             this.labelEUUserEmailNotEnabled.Visible = false;
             // 
+            // buttonEUCancel
+            // 
+            this.buttonEUCancel.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.buttonEUCancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
+            this.buttonEUCancel.Location = new System.Drawing.Point(221, 373);
+            this.buttonEUCancel.Name = "buttonEUCancel";
+            this.buttonEUCancel.Size = new System.Drawing.Size(88, 34);
+            this.buttonEUCancel.TabIndex = 21;
+            this.buttonEUCancel.Text = "Cancel";
+            this.buttonEUCancel.UseVisualStyleBackColor = true;
+            // 
+            // labelEUDisplayName
+            // 
+            this.labelEUDisplayName.AutoSize = true;
+            this.labelEUDisplayName.Location = new System.Drawing.Point(19, 74);
+            this.labelEUDisplayName.Name = "labelEUDisplayName";
+            this.labelEUDisplayName.Size = new System.Drawing.Size(72, 13);
+            this.labelEUDisplayName.TabIndex = 22;
+            this.labelEUDisplayName.Text = "Display Name";
+            // 
+            // textEUDisplayName
+            // 
+            this.textEUDisplayName.Location = new System.Drawing.Point(19, 90);
+            this.textEUDisplayName.Name = "textEUDisplayName";
+            this.textEUDisplayName.Size = new System.Drawing.Size(243, 20);
+            this.textEUDisplayName.TabIndex = 23;
+            this.textEUDisplayName.TextChanged += new System.EventHandler(this.textEUDisplayName_TextChanged);
+            // 
+            // errorEUNotify
+            // 
+            this.errorEUNotify.BlinkStyle = System.Windows.Forms.ErrorBlinkStyle.NeverBlink;
+            this.errorEUNotify.ContainerControl = this;
+            // 
             // formEditUser
             // 
-            this.AcceptButton = this.buttonEUApplyChanges;
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(415, 380);
-            this.Controls.Add(this.labelEUUserEmailNotEnabled);
+            this.CancelButton = this.buttonEUCancel;
+            this.ClientSize = new System.Drawing.Size(415, 419);
+            this.Controls.Add(this.textEUDisplayName);
+            this.Controls.Add(this.labelEUDisplayName);
+            this.Controls.Add(this.buttonEUCancel);
             this.Controls.Add(this.labelEUManagerName);
             this.Controls.Add(this.textEUManager);
             this.Controls.Add(this.labelEUManager);
@@ -265,9 +315,11 @@
             this.Controls.Add(this.labelEUPrimaryEmailAddress);
             this.Controls.Add(this.labelEUFirstName);
             this.Controls.Add(this.buttonEUApplyChanges);
+            this.Controls.Add(this.labelEUUserEmailNotEnabled);
             this.Name = "formEditUser";
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
             this.Text = "EditUser";
+            ((System.ComponentModel.ISupportInitialize)(this.errorEUNotify)).EndInit();
             this.ResumeLayout(false);
             this.PerformLayout();
 
@@ -294,5 +346,9 @@
         private System.Windows.Forms.TextBox textEUManager;
         private System.Windows.Forms.Label labelEUManagerName;
         private System.Windows.Forms.Label labelEUUserEmailNotEnabled;
+        private System.Windows.Forms.Button buttonEUCancel;
+        private System.Windows.Forms.Label labelEUDisplayName;
+        private System.Windows.Forms.TextBox textEUDisplayName;
+        private System.Windows.Forms.ErrorProvider errorEUNotify;
     }
 }