+++ PMC-Desktop/formEditUser.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -13,22 +14,27 @@ using System.Windows.Forms;
 namespace PMC_Desktop {
     public partial class formEditUser : Form {
         private static User UserObj = new User ();
+        private static User UserPrev = new User ();
         private static User ManagerObj = new User ();
+        private List<string> EmailList = new List<string> ();
         private static bool isFirstLoad = true;
 
         public formEditUser (User user) {
             InitializeComponent ();
 
             LoadUserInfo (user);
+            UserPrev.SetUser (user.Username, true);
         }
 
         public void LoadUserInfo (User user) {
             isFirstLoad = true;
             UserObj = user;
             Text = $"Editing: {user.userTools.DisplayName}";
+
             textEUFirstName.Text = user.userTools.GivenName;
             textEUMiddleName.Text = user.userTools.MiddleName ?? "";
             textEULastName.Text = user.userTools.Surname;
+            textEUDisplayName.Text = user.userTools.DisplayName;
 
             comboEUDepartment.SelectedIndex = comboEUDepartment.Items.IndexOf (user.Department);
 
@@ -40,12 +46,16 @@ namespace PMC_Desktop {
 
             PropertyValueCollection temp = user.UserObject.Properties["proxyaddresses"];
             List<string> proxyAddresses = new List<string> ();
-            foreach (var item in temp) {
-                proxyAddresses.Add (item.ToString ());
+            if (EmailList.Count > 0) {
+                proxyAddresses = EmailList;
+            } else {
+                foreach (var item in temp) {
+                    proxyAddresses.Add (item.ToString ());
+                }
+                string primary = proxyAddresses.FirstOrDefault (item => item.Contains ("SMTP:"));
+                proxyAddresses.Remove (primary);
+                proxyAddresses.Insert (0, primary);
             }
-            string primary = proxyAddresses.FirstOrDefault (item => item.Contains ("SMTP:"));
-            proxyAddresses.Remove (primary);
-            proxyAddresses.Insert (0, primary);
 
             comboEUPrimaryEmailAddress.Items.Clear ();
             listEUProxyAddresses.Items.Clear ();
@@ -68,6 +78,7 @@ namespace PMC_Desktop {
                 listEUProxyAddresses.Enabled = false;
                 buttonEUAddNewEmail.Enabled = false;
                 labelEUUserEmailNotEnabled.Visible = true;
+                labelEUUserEmailNotEnabled.BringToFront ();
             }
 
             isFirstLoad = false;
@@ -95,27 +106,104 @@ namespace PMC_Desktop {
 
         private void comboEUPrimaryEmailAddress_SelectedIndexChanged (object sender, EventArgs e) {
             if (!Regex.IsMatch (UserObj.Email, comboEUPrimaryEmailAddress.Text) && !isFirstLoad) {
-                UserObj.userTools.EmailAddress = comboEUPrimaryEmailAddress.Text;
-                List<string> list = new List<string> ();
                 foreach (object item in UserObj.UserObject.Properties["proxyaddresses"]) {
-                    list.Add (item.ToString ());
+                    EmailList.Add (item.ToString ());
                 }
-                string changeToSecondary = list.First (i => Regex.IsMatch (i, "SMTP"));
-                list.Remove (changeToSecondary);
-                string changeToPrimary = list.First (i => Regex.IsMatch (i, comboEUPrimaryEmailAddress.Text));
-                list.Remove (changeToPrimary);
+                string changeToSecondary = EmailList.First (i => Regex.IsMatch (i, "SMTP"));
+                EmailList.Remove (changeToSecondary);
+                string changeToPrimary = EmailList.First (i => Regex.IsMatch (i, comboEUPrimaryEmailAddress.Text));
+                EmailList.Remove (changeToPrimary);
                 changeToSecondary = changeToSecondary.Replace ("SMTP", "smtp");
                 changeToPrimary = changeToPrimary.Replace ("smtp", "SMTP");
-                list.Insert (0, changeToSecondary);
-                list.Insert (0, changeToPrimary);
-                UserObj.UserObject.Properties["proxyaddresses"].Clear ();
-                UserObj.UserObject.Properties["proxyaddresses"].AddRange (list.ToArray ());
+                EmailList.Insert (0, changeToSecondary);
+                EmailList.Insert (0, changeToPrimary);
                 LoadUserInfo (UserObj);
             }
         }
 
         private void buttonEUApplyChanges_Click (object sender, EventArgs e) {
+            UserObj.userTools.GivenName = textEUFirstName.Text;
+            UserObj.userTools.MiddleName = textEUMiddleName.Text;
+            UserObj.userTools.Surname = textEULastName.Text;
+            UserObj.userTools.DisplayName = textEUDisplayName.Text;
 
+            formConfirmEdit confirm = new formConfirmEdit (UserObj, UserPrev);
+
+            Visible = false;
+            confirm.ShowDialog ();
+            Visible = true;
+        }
+
+        private void textEUFirstName_TextChanged (object sender, EventArgs e) {
+            if (textEUFirstName.Text != UserObj.userTools.GivenName) {
+                EnableNotify (labelEUFirstName);
+            } else {
+                DisableNotify (labelEUFirstName);
+            }
+        }
+
+        private void EnableNotify (Control control, string text = "Edited") {
+            errorEUNotify.SetIconAlignment (control, ErrorIconAlignment.MiddleRight);
+            errorEUNotify.SetError (control, text);
+        }
+
+        private void DisableNotify (Control control) {
+            errorEUNotify.SetError (control, "");
+        }
+
+        private void textEUMiddleName_TextChanged (object sender, EventArgs e) {
+            if (textEUMiddleName.Text != UserObj.userTools.MiddleName) {
+                EnableNotify (labelEUMiddleName);
+            } else {
+                DisableNotify (labelEUMiddleName);
+            }
+        }
+
+        private void textEULastName_TextChanged (object sender, EventArgs e) {
+            if (textEULastName.Text != UserObj.userTools.Surname) {
+                EnableNotify (labelEULastName);
+            } else {
+                DisableNotify (labelEULastName);
+            }
+        }
+
+        private void textEUDisplayName_TextChanged (object sender, EventArgs e) {
+            if (textEUDisplayName.Text != UserObj.userTools.DisplayName) {
+                EnableNotify (labelEUDisplayName);
+            } else {
+                DisableNotify (labelEUDisplayName);
+            }
+        }
+
+        private void comboEUDepartment_SelectedIndexChanged (object sender, EventArgs e) {
+            if (comboEUDepartment.SelectedItem.ToString () != UserObj.Department) {
+                EnableNotify (labelEUDepartment);
+            } else {
+                DisableNotify (labelEUDepartment);
+            }
+        }
+
+        private void textEUTitle_TextChanged (object sender, EventArgs e) {
+            if (textEUTitle.Text != UserObj.Title) {
+                EnableNotify (labelEUTitle);
+            } else {
+                DisableNotify (labelEUTitle);
+            }
+        }
+
+        private void textEUManager_TextChanged (object sender, EventArgs e) {
+            var temp = ADS.GetSingleUser (textEUManager.Text);
+            if (temp != null) {
+                labelEUManagerName.Text = temp.Properties["displayname"][0].ToString ();
+                if (temp.Properties["distinguishedname"][0].ToString () != UserObj.UserObject.Properties["manager"][0].ToString ()) {
+                    EnableNotify (labelEUManager);
+                } else {
+                    DisableNotify (labelEUManager);
+                }
+            } else {
+                labelEUManagerName.Text = "User not found!";
+                EnableNotify (labelEUManager, "User not found!");
+            }
         }
     }
 }